app/pyproject.toml
[tool.poetry]
name = "pii-supplier-masking-app"
version = "0.1.0"
description = ""
authors = ["jbiddle <jbiddle96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyodbc = "^5.1.0"
marimo = "^0.3.2"
sqlalchemy = "^2.0.28"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
